Add Telegram bot to deliver telegram notifications
This bot requires that you set up a bot with Telegram. See the updated
README.md to see how.

It also, albeit only briefly, requires that you have apublically-
accessible domain name.

There are two methods of using it - with webhooks, if TSD is running on
a domain accessible from the internet, and without, if it is not.

If you're using a local tunnel and want to manually specify the domain
TSD will run on, replace `telegram_bot.py:24` with
`bot.set_webhook('https://example.com/telegram/`)`, replacing
`https://example.com/` with your local tunnel URL. This will enable full
interaction with the bot even when running the server locally.