Update unraid_guide.md (#659)
* Update unraid_guide.md

Change references to TSD to Obico

* Update unraid_guide.md

A few tweaks to change the reference to TSD in the old install back to what it likely is for someone who has installed it before. Additionally, checking out the repo to the folder "Obico" instead of "obico-server" if you don't specify the folder on git pull.

* Update unraid_guide.md

Updating the information to utilize the unRAID Plug-In Docker Compose Manager to get Docker Compose installed.