Update link to Jetson Nano sd card image
The current version of the sd card image linked to (4.5) does not match the version the script currently targets (4.6). 4.6.1 also cannot be used (as I painfully found out) because python appears to be upgraded to python3 (and I think the python3-is-python package is installed), so the docker-compose plugin does not get successfully installed and is unusable.

This confused me in multiple ways and I hope this changes helps anyone else setting up a Nano in the future.